#86 Rename module (parsers → asdata)
