+++ pandas_dataclasses/core/__init__.py
@@ -1,5 +1,5 @@
 # type: ignore
+from . import asdata
 from . import mixins
-from . import parsers
 from . import specs
 from . import typing
