+++ pandas_dataclasses/core/asdata.py
